+++ poc/classes/AuxSTConstants.py
@@ -1,0 +1,2 @@
+class AuxSTConstants:
+    qualified_re = r"\[([0-9,\+\*\[\]:a-zA-Z_\$]*)\]"