+++ poc/classes/AuxSTPredicateInstance.py
@@ -12,3 +12,16 @@ class AuxSTPredicateInstance(AuxSTInstance):
         self.zto = i.last_positions_by_rule['PredicateInstance'].pos_to_str()
         self.keyword = ""
 
+    def clone(self):
+        new_predicate = self._copy(AuxSTPredicateInstance(self._i))
+        new_predicate.id = self.id
+        new_predicate.def_type = self.def_type
+        new_predicate.keyword = self.keyword
+        new_predicate.zfrom = self.zfrom
+        new_predicate.to = self.zto
+        new_predicate.mandatory = self.mandatory
+        new_predicate._is_inherited = True
+        return new_predicate
+
+
+