+++ poc/classes/AuxSTVarSpecList.py
@@ -7,3 +7,5 @@ class AuxSTVarSpecList(AuxSTOutline):
     def __init__(self):
         super().__init__(None, AuxSymbolTable.var_spec)
 
+    def clone(self):
+        return AuxSTVarSpecList()